+++ Sources/Datadog/Logger.swift
@@ -49,6 +49,24 @@ public protocol LoggerProtocol {
     /// the same key already exist in this logger, it will be overridden (only for this message).
     func log(level: LogLevel, message: String, error: Error?, attributes: [String: Encodable]?)
 
+    /// General purpose logging method.
+    /// Sends a log with certain `level`, `message`, `errorKind`,  `errorMessage`,  `stackTrace` and `attributes`.
+    ///
+    /// This method is meant for non-native or cross platform frameworks (such as React Native of Flutter) to send error information
+    /// to Datadog. Although it can be used directly, it is recommended to use other methods declared on `Logger`.
+    ///
+    /// - Parameters:
+    ///   - level: the log level
+    ///   - message: the message to be logged
+    ///   - errorKind: the kind of error reported
+    ///   - errorMessage: the message attached to the error
+    ///   - stackTrace: a string representation of the error's stack trace
+    ///   - attributes: a dictionary of attributes (optional) to add for this message. If an attribute with
+    /// the same key already exist in this logger, it will be overridden (only for this message).
+    func log(level: LogLevel, message: String, errorKind: String?, errorMessage: String?,
+             stackTrace: String?, attributes: [String: Encodable]?)
+
+
     // MARK: - Attributes
 
     /// Adds a custom attribute to all future logs sent by this logger.
@@ -184,6 +202,12 @@ public class Logger: LoggerProtocol {
 
     public func log(level: LogLevel, message: String, error: Error?, attributes: [String: Encodable]?) {
         v2Logger.log(level: level, message: message, error: error, attributes: attributes)
+    }
+
+    public func log(level: LogLevel, message: String, errorKind: String?, errorMessage: String?,
+             stackTrace: String?, attributes: [String: Encodable]?) {
+        v2Logger.log(level: level, message: message, errorKind: errorKind, errorMessage: errorMessage,
+                     stackTrace: stackTrace, attributes: attributes)
     }
 
     public func addAttribute(forKey key: AttributeKey, value: AttributeValue) {
@@ -421,6 +445,12 @@ internal struct CombinedLogger: LoggerProtocol {
         combinedLoggers.forEach { $0.log(level: level, message: message, error: error, attributes: attributes) }
     }
 
+    func log(level: LogLevel, message: String, errorKind: String?, errorMessage: String?,
+             stackTrace: String?, attributes: [String : Encodable]?) {
+        combinedLoggers.forEach { $0.log(level: level, message: message, errorKind: errorKind, errorMessage: errorMessage,
+                                         stackTrace: stackTrace, attributes: attributes) }
+    }
+
     func addAttribute(forKey key: AttributeKey, value: AttributeValue) {
         combinedLoggers.forEach { $0.addAttribute(forKey: key, value: value) }
     }
@@ -448,6 +478,7 @@ internal struct CombinedLogger: LoggerProtocol {
 
 internal struct NOPLogger: LoggerProtocol {
     func log(level: LogLevel, message: String, error: Error?, attributes: [String: Encodable]?) {}
+    func log(level: LogLevel, message: String, errorKind: String?, errorMessage: String?, stackTrace: String?, attributes: [String : Encodable]?) {}
     func addAttribute(forKey key: AttributeKey, value: AttributeValue) {}
     func removeAttribute(forKey key: AttributeKey) {}
     func addTag(withKey key: String, value: String) {}